requirements.txt
# Core Django Framework
Django==5.2.4
djangorestframework==3.16.0
python-dotenv==1.1.1

# Database
psycopg2-binary==2.9.10

# Image Processing & Media
Pillow==11.3.0

# Excel Processing
openpyxl==3.1.2

# API Documentation
drf-yasg==1.21.10

# Authentication & Security
django-allauth==65.10.0
djoser==2.3.3
django-otp==1.6.1
django-two-factor-auth==1.17.0
django-phonenumber-field==8.1.0
django-rest-passwordreset==1.5.0
djangorestframework-simplejwt==5.5.0
pyotp==2.9.0
qrcode==7.4.2

# Filtering & Search
django-filter==25.1

# CORS & Security
django-cors-headers==4.7.0
django-ratelimit==4.1.0
django-axes==6.1.1
django-csp==3.7

# Form Styling
django-crispy-forms==2.3
crispy-bootstrap4==2022.1

# Production & Deployment
gunicorn==23.0.0
whitenoise==6.9.0

# Cloud Storage & Media
django-storages==1.14.6
boto3==1.40.45
cloudinary==1.44.1

# Task Queue & Background Jobs
celery==5.5.3
django-celery-beat==2.8.1
django-celery-results==2.6.0
redis==7.0.0
django-redis==6.0.0

# Utilities & Dependencies
requests==2.32.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
django-formtools==2.5.1
django-timezone-field==7.1

# Social Authentication
social-auth-app-django==5.5.1
social-auth-core==4.7.0

# Additional Dependencies
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
botocore==1.40.45
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==2.0.6
cryptography==45.0.5
defusedxml==0.7.1
django-daisy==1.1.2
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
prompt-toolkit==3.0.52
pycparser==2.22
PyJWT==2.9.0
pypng==0.20220715.0
python-crontab==3.3.0
python3-openid==3.2.0
requests-oauthlib==2.0.0
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14